+++ packages/app/src/common/components/Toolbar/style.scss
